Added experimental timing code to the X11 and SDL versions. Now its possible to use the emulator without hogging the CPU.  Usage on most machines I've tested is around 1% - 2%, versus the 100% that was used before.
SDL version seems to like this new code better.  I think that usleep is
causing some problems in the X11 version.  I'd appreciate some feedback
on this.  Maybe I'll change to select or something.

There is a slight skip in framerate that can't be avoided unfortunately.
You may not even notice it.  Sound may be slightly different as well.  In
my opinion, this is a good tradeoff for basically reducing CPU usage to nil.