More code changes and reorganization for key remapping.  The SDL version now uses the EventHandler class and doesn't directly handle events itself.
All thats left for key remapping is to create load/save methods for the
event arrays.  Of course, to actually change events from within the
emulator will require a GUI, which I haven't even started yet :)  The GUI
will have to wait until the graphics rewrite, where we go to a full
320x200 framebuffer (versus the 160x200 one we have now).

In the meantime, remapping will be configurable from the rcfile only.