This is the first pass at getting the Cyberstella port working. It now partially compiles again, but many features are still missing.
Added a FrameBufferWin32 class.  This will eventually contain
DirectDraw windowed and fullscreen code, but for now it does
nothing.

Upgraded the project to Visual Studio .NET.  It will no longer
compile with VC++ 6.

Now for the parts that will definitely change:

1) Right now, the MainWin32 class is also acting as the FrameBuffer.
The relevant code will be moved to FrameBufferWin32.

2) The DirectInput class will be completely rewritten to use
DirectInput8.  Also, I'll be moving to event-based input where
the state is taken into account.  The rewrite will also allow
multiple joysticks, and selecting which joystick input device
is mapped to which internal Atari joystick.

3) The SettingsWin32 class does nothing ATM.  It will be expanded,
and will replace using the registry (the associated registry class
will be removed).

4) I haven't even looked at the sound yet.  I want to get all
other stuff working before I even attempt it.  It *did* make
some noise the last time I tried it, but something wasn't quite
right.

So, bottom line is that this compiles without sound, mouse,
joystick, windowed modes, and any of the 1.4 core enhancements.
But at least it compiles :)  I expect the extras to be worked
out quite soon (except for sound), since I've already done
most of the same thing in Linux.

And to potential porters; its still not ready, and you should
probably be looking at the SDL port anyway ...