First of all, I forgot to include the MainWin32 class before, so it really didn't compile :)
And now for the changes.

1) I've updated the DirectInput class to DirectInput 8, and moved
to an event-based system, somewhat similar to SDL.  Only the
keyboard is supported for now, but adding joystick and mouse should
be quite easy.

2) I've (almost completely) moved the FrameBuffer stuff out of
MainWin32 and into FrameBufferWin32, where it actually belongs.

Next, I'll work on more code cleanups, and then I'll attempt
to switch to the experimental DirectDraw code (so we have
windowed *and* fullscreen modes dynamically).

All things considered, its coming together faster than I thought ...