Some more cleanups wrt the sharing of the SDL codebase between Linux and Windows.
Reworked the Makefile (again).  Now you have to edit it and select
which system you are using for the SDL port (linux or windows),
then start compilation with 'make sdl'.

The makefile is starting to get harder to manage, and with the
explosion of features, I'm looking into moving to a configure
script (autoconf and automake) for a future release.