Made the Driving class actually refer to Driving Controller events instead of joystick events.
This bug has been present in the codebase for over 3 years (at least).
Thanks to Mark Grebe for pointing it out.

This means that you will no longer be able to use a gamepad/joystick
for both joystick *and* driving controller events at the same time;
you'll have to map a particular device to either an internal Atari
joystick *or* an internal Atari Driving Controller.

This also means that the Driving Controller remapping now actually
works.  Previously, it didn't do anything.