Removed dependency of Console from the Settings class.  It didn't make sense for the Settings class to have to depend on Console, especially since I plan to use the Settings class in the Windows frontend.
Removed the snapshot filename code from the individual SettingsXXX classes,
since it was redundant.  Instead, added a ::fileExists() method, which
is the only thing that was really platform-specific about the code
anyway.

Linux and Windows ports have been updated.  The Mac port will have to be
updated to work with this.