Removed the virtual Settings::usage() method from the SettingsUNIX and SettingsWin32 classes, since they both did the same thing, and placed the method in the Settings class itself (ifdef'ed out for OSX).  It was done this way when core was still in flux and each port didn't support all the same options.  Since the core is now unified, it didn't make sense to have a separate usage() method for each port.
Added the '-listrominfo' argument (can only be called from the commandline,
not set with the INI file).  This returns selected info from the stella.pro
file and exits immediately.  It returns data in the form of:

   # of roms (first line)
   MD5|Name|Rarity|Manufacturer|Note (# of roms lines) ...

This makes it much easier for GUI frontends to interface with Stella,
since all stella.pro parsing is now done within Stella itself.

Added various print() methods to Properties and PropertiesSet to
facilitate the above '-listrominfo' argument.

Fixed a bug in PropertiesSet::size(); it always returned 0.