Cleaned up the SDL event gathering loop.  It now almost exclusively makes use of switch/case statements instead of the slower if/elseif structure.  Hopefully this will fix event problems some users were experiencing.
The Control or Alt button can actually now be used in event remapping.
Due to the old event handling, the Control key could be assigned to
an event but would never actually be usable.

Removed the option to remap the exit event (for now), since I haven't
figured out a way to remap it *back* to 'Escape' if that mapping is
ever removed.

Added beginning of support for Stella to automatically determine the
correct framerate based on ROM format type.  The final version of this
will probably wait until Stella 1.5.