Added Game Info Dialog.  It basically lists all the relevant information in the stella.pro file for the particular game.
Added Help Dialog, which gives a description of the most commonly
used non-remappable commands.  This still has to be adapted for OSX
shortcuts.

Finished the Sound Dialog.  Still TODO is remove the requirement for
restarting Stella before sound can be disabled.

All that's left now are the Miscellaneous and Event Remapping Dialogs.
The later will be the hardest, but it will also work a lot better than
in previous versions of Stella.