More work on the EventHandler and EventMappingDialog classes.  The EventHandler class now contains string mappings for all possible SDL keys, which should hopefully take care of our international users.
Added 'Map', 'Erase' and 'Cancel' buttons to the EventMappingDialog.
They still don't do much, but it will now be an easy matter to add the
hooks into EventHandler.  You *are* able to scroll through the list
and see which keys are bound to which actions, though.

This approach is *much* cleaner than in previous code.  Now the EventHandler
actually does the remapping; the EventMappingDialog only collects the
events and associated actions and passes them on.  This is much cleaner
than before, when (for example) the FrameBuffer was doing remapping!  What
the hell did the framebuffer have to do with event mapping???