Implemented BrowserDialog functionality.  That means we can now browse for the correct ROM folder, and the folders contents will be shown in the listview.
Still TODO is modify GameList so that it contains all relevant info for
a ROM.  Then, the listview will contain the actual names of the ROMS
(as specified in stella.pro) instead of their on-disk filenames.  And at
that point, we can actually start a game directly from that listview.