Decided not to add a MiscDialog box, since there weren't really any options to add to it.
Reworked the EventMappingDialog to take up less screen real-estate, and added
the "Mouse is paddle x" option there (since technically it's also a mapping).

Added an 'emphasized' text color to the list of available GUI colors, and
made some text in the EventMappingDialog use that color.

Added AboutDialog, which lists info such as version of Stella, supported
features, list of developers and contributors, and thanks to certain groups.