Changed some FIXME's to TODO's, since they probably won't be done for the next release.
Added support for inspecting the modification time of the rom directory,
and forcing a reload if that directory has been changed.  So, if you add
or delete a ROM from the romdir, the next time you start Stella, it will
automatically reload the rom listing for you.  Pretty cool I think :)

Added support for each derived OSystem to set its video drivers, and support
to the VideoDialog to show them in a dropdown menu.  That means that Win32
users will be able to select between 'windib' and 'directx' in the VideoDialog
(when I update the Win32 code to do so).