Added first pass of CpuWidget.
Renamed ByteGridWidget to DataGridWidget, because it can be used to accept data
other than bytes (also words).

Further work on making the GUI use whichever base you set as default.  This
one seems to be a hard nut to crack.

Changed behaviour of '-debugheight' to specify the number of lines to use
for the debugging area, vs the absolute pixels to use.  So '-debugheight 20'
would make the debugger be 20 lines tall (in the PromptWidget).  The minimum
height is now 15 lines, and all GUI elements will be drawn according to this.

Still TODO is get the 'height' command to work in the debugger.