Added [] operator to expression syntax. foo[bar] is exactly equivalent to *(foo+bar), just like it is in C.
Added _bank special "variable" to expression syntax. It returns the current
bank, so you can set bankswitching breakpoints:

	breakif {_bank==1 && pc==whatever}

It looks like I'm going to be using _ as a prefix that means "special
meaning; reserved; don't use in your program", just like C does. Some
existing games might need to be rewritten if they use _bank or _scan as
a label... but it'll be well worth the price, to the game author.