OK, I'm not sitting on this any longer, for fear it'll get too far out of sync.  Implemented dirty widget support for the GUI.  That means the widgets will only be drawn when necessary.  There are still a few gotcha's:
1)  OpenGL mode hasn't been ported to this new scheme.
2)  It's not totally finished, so some artifacts may appear onscreen.
3)  Selecting active widgets with the mouse is borked.
4)  Prompt commands that change the core aren't shown in the other
    tabs/interface.  Fixing this will require some infrastructure work
    in Debugger and DebuggerParser.
5)  A lot of print debug code has been left in; please ignore it for now.

Moved a lot of the debugger widgets to use non-proportional font (still
TODO is get a larger font) and not use 'magic numbers' for the layout.
That means when a new font is added, the layout should re-arrange itself.

Moved various Debugger tab widgets from 'src/gui' to 'src/debugger',
because they shouldn't be compiled when debugger support isn't
included.  So now (for example), RamWidget and RamDebug are both in
the debugger directory.

Probably more stuff I'm forgetting about.  It looks like the ScummVM
code can be made adequately fast, so the jump to Qt won't be necessary.