More efficiency fixes.  Redrawing a tab (in the TabDialog) no longer forces a redraw of the whole parent dialog (this is a big optimization in the case of the DebuggerDialog, since it will soon have a resolution of 1024x768).
Made the TiaOutputWidget act as a dirty rect, and only redraw itself
when necessary.

Fixed selection of items in TabWidget which have the
WIDGET_TAB_NAVIGATE property.  Clicking them with the mouse now
correctly selects them (and deselects other widgets in that chain).
Still TODO here is have the TabWidget draw the outline around such
widgets itself, so that when one is selected and the others
deselected, only the outlines need to be redrawn (vs. the whole tab).

Removed FrameBuffer::blendRect(), since the GUI code will never
have blended rectangles (it would cause to much complexity in the
dirty update/rectangle code).