Added Undo, Revert, Search, Compare and Reset buttons to the RamWidget. The latter three completely replace the functionality of the CheatWidget, so it will soon disappear.
The results of a RAM search are indicated by a blue frame around a cell.
Compare still isn't working, but I know how to fix it.

Made EditableWidget a CommandSender, and have it send signals when
data entry is complete or cancelled.

Some API cleanups in FrameBuffer/DialogContainer wrt refreshOverlay()
and refreshTIA().