Broke old ListWidget down to a base class, and created a new StringListWidget and CheckListWidget to extend from it.  This means the derived classes can concentrate on their main purpose, and leave all the ListWidget stuff to the parent class.
Turned the TIA debugger area into an absolute size; 320x260.  This
should be large enough to accommodate all PAL ROMS.  This means the
tabset got shorter, but it's just too difficult to add widgets when
the area can change.  A zoom widget is going to be added to the right
of the TIA image, and we need absolute sizes for that area.
Besides, now that area is usable for other widgets ...