Added mini message box to debugger, which will be used to inform the user of debugger-related state information.
Reworked breakpoint/trap support to display a message in debugger
messagebox indicating what condition caused the debugger to start.
This is much better than before, when no feedback was given as to
why the debugger started.

Some usability fixes to ListWidget/RomWidget.  There are still a few
more tweaks to do before the final release.