Updating the PC in CpuWidget area now moves the PC indicator in ROM area.
Disabled editing of banks when the ROM doesn't do bankswitching.

Fixed Shift-Tab key not being detected (and hence not being able to
reverse-tab through items in debugger).

Bumped version number to 2.0 and updated all documentation.  Barring
anything major popping up in the next day or so, this is essentially the
2.0 release.