Fixed 'debugheight' argument and associated 'height' debugger command to correctly set the height of the Prompt/tab area in the debugger. The new minimum height is 27 lines in the prompt, which corresponds to an ~ 700 pixel high window.  This should fix the problems OSX users were having with the dock taking up too much space onscreen in 1024x768 mode.
Changing the height from within Stella still requires a reload of the ROM,
and now a message is printed stating this fact.  The GUI code is in need
of a minor overhaul so that resizing can be done automatically, without
requiring a ROM reload and also taking font size into account.  This will
be completed for a future release.

Finally, the RomWidget is now also resized when the debugger height
changes, taking the extra vertical space into account.