Reworked the mapping of joysticks yet again.  After some thought, it made no sense to restrict Stella to only work with two joysticks, since there's no reason one couldn't use four devices, with each one emulating a paddle (once I get paddle remapping done).  So I've removed that restriction, and also removed the 'leftport' and 'rightport' arguments.
The one case where remapping a joystick *does* make sense is for
multiple Stelladaptors.  In that case, there might only be one
Stelladaptor installed, but you'd like it to emulate the right virtual
port.  Previously this wasn't possible, but now it is with the addition
of the 'sa1' and 'sa2' commandline arguments.  These arguments accept
either 'left' or 'right', specifying which virtual port to emulate.

Updated the InputDialog virtual device tab for remapping Stelladaptors.

Shortened some keynames in EventHandler, so that the actual character is
used (for example, ? instead of QUESTION).

Some fixes for the OS/2 port.