Reworked overlaid messages in the framebuffer; you can now select the region of the screen in a 9x9 grid (top/middle/bottom by left/center/right), as well as the color of the message text.
Updated some strings on the ROM launcher, since it doesn't make sense
to 'Play' a directory.

Disabled display of 'Note' while in rom browse mode in the ROM launcher.
It ends up taking too much CPU time to parse all the ROMs in a directory
each time it changes, so there are now two options: (1) don't browse the
filesystem, and precompute the folder contents/see all the ROM info
or (2) browse the filesystem, allowing selection of folders, etc, but
not seeing ROM names and notes as defined in the properties file.  For
performance reasons, it's going to have to be an either/or thing.