Tweaked cart autodetection code wrt '3E' bankswitching.  Since a '3E' cart is almost always a '3F' cart as well, it makes sense to only test for it when we already know the cart is probably '3F'.  As well, the pattern for a '3F' cart results in a lot less false-positives than '3E'.
Added '-colorloss' commandline argument and 'Ctrl-l' key shortcut.  This
enables or disables the PAL color-loss effect, and will default to off
(similar to z26 functionality).  Still TODO is add a UI entry for it.

Fixed a bug in the calculation of PAL color-loss values; it wasn't actually
giving grayscale data, but something with a greenish tint.

Updated PAL autodetection code to run for 60 test frames instead of 20,
and ignore the first 30.  This helps when loading SuperCharger games,
since the BIOS code always runs at NTSC rates, even if the ROM really
should be PAL.  For now, this means it only works when 'fastscbios' is
enabled, and the BIOS runs in 30 frames or less (vs. needing 250+ frames
for normal speed!).  I'm strongly leaning to just making 'fastscbios'
a non-editable default.