OK, some huge changes across the board, so lets see if I get it all:
After much request, added ability to access the settings menu from the
ROM browser dialog.  This menu now contains almost all items that can
be selected in Stella, and can be accessed in-game as before.

Completely removed pause functionality from the core code.  It made
sense back when Stella was a single-mode program: there were two modes;
emulation and pause.  Now that there are other event modes, the
EventHandler state machine is getting too complicated.  If you want to
pause, you can simply enter one of the in-game menus.  Related to this,
when the app is minimized, Stella enters the menu dialog state.
Previously, minimizing the app caused a pause, but since there was no
onscreen feedback, many people assumed the app locked up.

Added centering to all Dialog boxes, which is done dynamically, as they're
placed on the dialog stack to be drawn to the screen.

Cleaned up the API of Console/FrameBuffer/OSystem classes wrt to palettes
and timing.  Parts of each were being done in different classes; now it
should be more consistent.

Started infrastructure for user-selectable UI palettes.  For now, there's
no way to change it in the GUI, and it defaults to the normal palette.
Eventually, there will be several choices selectable from an in-game
menu.

Removed '-channels' commandline argument, since that feature can be
set from the ROM properties.

Added '128' to the choices for fragment size in AudioDialog.

Tweaked the OpenGL dynamic loading code to test both the given GL
lib, and if that fails to use auto-detection.  It seems in the OSX port,
the first approach works for some people, and not the other (and vice-versa),