Relaxed recent equate changes to allow all symbols to be used from a .sym file.  This will have to do for now, until we find a better way of differentiating between constants and addresses.
Made the ROM disassembly area take advantage of a wide debugger window
by spacing out the label and disassembly areas.

Finalized location of the AVox and SaveKey EEPROM files; they're now
located in the basedir (the actual location depends on the OS).

Made deadzone for analog joysticks configurable from the commandline
and UI.  Added '-joydeadzone' commandline argument, which accepts
a value from 0 - 29, specifying a deadzone of '3200 + DEADZONE * 1000'.
Also added UI to InputDialog to set this from dynamically from the UI.

Added code to only save the AVox and SaveKey EEPROM data file when
necessary.  This is for those systems based on flash storage, where
unnecessary writes will wear down the drive.

Added SaveKey as a full-fledged controller to GameInfoDialog.