OK, another large FrameBuffer-related change so that fullscreen, double-buffered modes don't have flicker.  Basically, the core FrameBuffer class has been modified to draw dialogs multiple times when necessary, above and beyond when changes are made (ie, if a new surface is to be overlaid, all surfaces underneath must be drawn to BOTH buffers in double-buffered mode first).
Software rendering is currently broken in fullscreen mode, but I know
why.  It's basically the optimizations I try to make for large software
surfaces (the 'isBase' functionality).  This is going to disappear before
the next release.  And in case I haven't mentioned it for a while; I
REALLY HATE software mode.