Baring any last minute issues, it looks like I've finally finished with the new FrameBuffer surface-based instrastructure.  Thank God, since it was really starting to get on my nerves :(
Made the Alt- and Alt= keys (for changing videomodes) only be active
in TIA emulation mode.  This solves the final remaining positioning
issues with ContextMenus, since now a ContextMenu is never onscreen when
the VideoMode changes size (and hence we don't have to worry about
re-positioning it).

Now to fix the RomInfoWidget, and then UI-wise we'll have parity with
the 2.6.1 release ...