K, now that release 2.7 is out of the way, I can move on again.
First pass at cleaning up the TIA class, in anticipation of the new
TIA infrastructure with improved HMOVE emulation.  This code has been
in need of a cleanup for some time now; there are commented TODO's
and FIXME's that are up to 10 years old!

Moved all static constants and tables into a separate TIATables class,
making the main TIA class much shorter and easier to understand.

TIA no longer uses or honours the 'Allow HMOVE blanks' ROM property;
further work will remove this from the codebase entirely.

Removed MediaSource abstraction and have the codebase use the TIA
directly.  It was probably a good idea to have this abstraction when
Stella was first written (in anticipation of multiple TIA
implementations), but seeing how that hasn't happened for approx.
13 years, I don't see it ever happening :)  Besides, we're paying a
price for that abstraction (certain functions can't be inline'd) for
no real reason.  Perhaps in the future, the M6502/M6502Low/M6502Hi
will experience a similar contraction.