Added '-gl_aspectn' and 'gl_aspectp' commandline arguments (and associated UI items) to set the aspect ratio to use in NTSC and PAL modes.  This replaces the '-gl_aspect' setting.
Hold-select/reset/button0 now work from the UI.

Grabmouse can now be set in the UI.

Changed status of the fix for ATI/OpenGL/sound from fixed to workaround.
Sound will work for the first ROM, but not for subsequent ones.  This
improves on the previous behaviour, where sound didn't work at all.

Bumped version to 2.7.1.