OK, this is going to be a BIG commit.  Changed all source files to process the $Id$ inline variable when doing an svn commit, similar to how CVS worked.
Removed rsynth code and associated Speakjet emulation functionality.
It hasn't been worked on in quite some time, and we never did fully
decide how to do it correctly.  At this point, I don't have the time
(or background) to continue work on it.  If someone steps up to
further continue its development, it can always be re-introduced
again.